Fix activity template parameters max and min occurences.
git-svn-id: http://forge2.pole-aquinetic.fr/svnroot/fesapi/trunk@87 553bd026-d4ad-425d-a02a-e125c7656c9a

+++ src/resqml2_0_1/ActivityTemplate.h
@@ -94,8 +94,8 @@ namespace resqml2_0_1
 		const bool & getParameterIsOutput(const std::string & paramTitle) const;
 		const unsigned int getParameterMinOccurences(const unsigned int & index) const;
 		const unsigned int getParameterMinOccurences(const std::string & paramTitle) const;
-		const unsigned int getParameterMaxOccurences(const unsigned int & index) const;
-		const unsigned int getParameterMaxOccurences(const std::string & paramTitle) const;
+		const long long getParameterMaxOccurences(const unsigned int & index) const;
+		const long long getParameterMaxOccurences(const std::string & paramTitle) const;
 
 		const std::vector<Activity*> & getActivitySet() const {return activitySet;}
